chore(functions): Protecting all endpoints; moving export and import to callabls;

+++ functions/src/callable/remove-user.ts
@@ -1,7 +1,14 @@
 import * as functions from 'firebase-functions';
 import {auth} from 'firebase-admin';
 
-export const removeUser = functions.https.onCall(async data => {
+export const removeUser = functions.https.onCall(async (data, context) => {
+  if (!context.auth) {
+    throw new functions.https.HttpsError(
+      'failed-precondition',
+      'The function must be called ' + 'while authenticated.'
+    );
+  }
+
   try {
     await auth().deleteUser(data.id);
   } catch (e) {
